detach site_filter only when in /admin

+++ controllers/components/multisite.php
@@ -45,7 +45,7 @@ class MultisiteComponent extends Object {
 	}
 
 	function initialize(&$controller, $settings = array()) {
-		if (1 == $controller->Auth->user('role_id')) {
+		if (1 == $controller->Auth->user('role_id') && isset($controller->params['admin'])) {
 			$controller->Node->Behaviors->detach('Sites.SiteFilter');
 		}
 		$site = Sites::currentSite();
